refactor: move loanCallouts to own file
