+++ src/components/LoanCards/KivaClassicBasicLoanCardExp.vue
@@ -174,7 +174,8 @@ import numeral from 'numeral';
 import { mdiChevronRight, mdiMapMarker, mdiCheckCircleOutline } from '@mdi/js';
 import { gql } from '@apollo/client';
 import * as Sentry from '@sentry/vue';
-import { readLoanFragment, watchLoanData, loanCallouts } from '@/util/loanUtils';
+import loanCallouts from '@/util/loanCallouts';
+import { readLoanFragment, watchLoanData } from '@/util/loanUtils';
 import { createIntersectionObserver } from '@/util/observerUtils';
 import LoanUse from '@/components/LoanCards/LoanUse';
 import percentRaisedMixin from '@/plugins/loan/percent-raised-mixin';
